refactor: moved function to own files
