+++ src/lib/is-empty-object.ts
@@ -1,0 +1,2 @@
+export const isEmptyObj = (obj: object): boolean =>
+  Object.keys(obj).length === 0;
